backend user, frontent and mobile

+++ Backend/SRC/controllers/notesController.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from "../models/note";
+import { assertIsDefined } from "../util/assertIsDefined";
 
-export const getNotes: RequestHandler = async (_req, res, next) => {
+export const getNotes: RequestHandler = async (req, res, next) => {
+
+  const authenticatedUserId = req.session.userId
   try {
-   
-    const notes = await NoteModel.find().exec();
+   assertIsDefined(authenticatedUserId)
+
+    const notes = await NoteModel.find({userId:authenticatedUserId}).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -15,13 +19,18 @@ export const getNotes: RequestHandler = async (_req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId
   try {
+    assertIsDefined(authenticatedUserId)
     if (!mongoose.isValidObjectId(noteId)){
         throw createHttpError(400, "keine gültige Notiz ID")
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "die Notiz nicht gefunden");
+    }
+    if(!note.userId.equals(authenticatedUserId)){
+      throw createHttpError(401,"401 error")
     }
 
     res.status(200).json(note);
@@ -42,13 +51,16 @@ export const createNote: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId
 
   try {
+    assertIsDefined(authenticatedUserId)
     if (!title) {
       throw createHttpError(400, "die Notiz benötigt einen Title");
     }
 
     const newNote = await NoteModel.create({
+      userId:authenticatedUserId,
       title,
       text,
     });
@@ -71,7 +83,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId=req.params.noteId
     const newTitle=req.body.title
     const newText=req.body.text
+    const authenticatedUserId = req.session.userId
     try {
+      assertIsDefined(authenticatedUserId)
         if (!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "keine gültige Notiz ID")
         }
@@ -82,6 +96,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
           if (!note) {
             throw createHttpError(404, "die Notiz nicht gefunden");
           }
+        if(!note.userId.equals(authenticatedUserId)){
+          throw createHttpError(404,"404")
+        }
           note.title=newTitle
           note.text=newText
           const updatedNote =await note.save()
@@ -94,7 +111,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler=async(req, res, next)=>{
     const noteId=req.params.noteId
+    const authenticatedUserId = req.session.userId
     try {
+      assertIsDefined(authenticatedUserId)
         if (!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "keine gültige Notiz ID")
         }
@@ -103,6 +122,9 @@ export const deleteNote: RequestHandler=async(req, res, next)=>{
         if(!note){
             throw createHttpError(404, "Notitz nicht gefunden")
 
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+          throw createHttpError(404,"404")
         }
 
         await NoteModel.findByIdAndRemove(noteId).exec()
